refactor: move siteData to a portable feature module and expose it via package exports
